Renamed slurm support
